+++ .eslintrc.js
@@ -1,13 +1,14 @@
 module.exports = {
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es2021": true
+    env: {
+        browser: true,
+        commonjs: true,
+        es2021: true,
     },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "ecmaVersion": 12
+    extends: "eslint:recommended",
+    parserOptions: {
+        ecmaVersion: 12,
     },
-    "rules": {
-    }
+    rules: {
+        indent: ["warn", 4],
+    },
 };
