+++ actions/craftingMaterials1-1.js
@@ -7,18 +7,18 @@ const { provider, nonceManager } = require("../config/wallet");
 const { log, error } = require("./utils");
 
 const contract = new ethers.Contract(
-  craftingContractAddress,
-  rarityAbi,
-  provider
+    craftingContractAddress,
+    rarityAbi,
+    provider
 );
 const writeContract = contract.connect(nonceManager);
 const { pause } = require("./utils");
 
 // get the time until next adventure
 const getAdventureLog = async (id) => {
-  await provider.ready;
-  let adventurersLog = await contract.adventurers_log(id);
-  return adventurersLog;
+    await provider.ready;
+    let adventurersLog = await contract.adventurers_log(id);
+    return adventurersLog;
 };
 
 // sends your summoner on an adventure!
@@ -26,20 +26,20 @@ const craftAdventure = async (summonerId, currentTime) => {
     let adventureTimestamp = await getAdventureLog(summonerId);
     let rewards = await contract.scout(summonerId);
     if (
-      currentTime.gt(adventureTimestamp) &&
+        currentTime.gt(adventureTimestamp) &&
       rewards.gt(ethers.BigNumber.from(0))
     ) {
-      try {
-        let response = await writeContract.adventure(summonerId);
-        /* let receipt = */ await response.wait();
-        // log("craftAdventure", summonerId, receipt);
-        log("craftAdventure", summonerId, `Adventure successfull!`);
-        pause();
-      } catch (err) {
-        error("craftAdventure", summonerId, `Could not send the tx: ${err}`);
-      }
+        try {
+            let response = await writeContract.adventure(summonerId);
+            /* let receipt = */ await response.wait();
+            // log("craftAdventure", summonerId, receipt);
+            log("craftAdventure", summonerId, `Adventure successfull!`);
+            pause();
+        } catch (err) {
+            error("craftAdventure", summonerId, `Could not send the tx: ${err}`);
+        }
     } else {
-      log("craftAdventure", summonerId, `Not yet time to adventure.`);
+        log("craftAdventure", summonerId, `Not yet time to adventure.`);
     }
 };
