+++ actions/gold.js
@@ -10,17 +10,17 @@ const contract = new ethers.Contract(goldContractAddress, rarityAbi, provider);
 const writeContract = contract.connect(nonceManager);
 
 const claimable = async (id) => {
-  let amount = contract.claimable(id);
-  return amount;
+    let amount = contract.claimable(id);
+    return amount;
 };
 
 const claim = async (summonerId) => {
     try {
-      let response = await writeContract.claim(summonerId);
-      /* let receipt = */ await response.wait();
-      // log("gold", summonerId, receipt);
-      log("gold", summonerId, `Gold claimed successfully!`);
+        let response = await writeContract.claim(summonerId);
+        /* let receipt = */ await response.wait();
+        // log("gold", summonerId, receipt);
+        log("gold", summonerId, `Gold claimed successfully!`);
     } catch (err) {
-      error("gold", summonerId, err);
+        error("gold", summonerId, err);
     }
 };
