+++ actions/mint.js
@@ -6,9 +6,9 @@ const rarityAbi = require("../abis/rarity.json");
 const { provider, nonceManager, wallet } = require("../config/wallet");
 
 const contract = new ethers.Contract(
-  rarityContractAddress,
-  rarityAbi,
-  provider
+    rarityContractAddress,
+    rarityAbi,
+    provider
 );
 
 const { pause } = require("./utils");
@@ -17,27 +17,27 @@ const { pause } = require("./utils");
 const writeContract = contract.connect(nonceManager);
 
 const mint = async (type) => {
-  let nonce = await wallet.getTransactionCount();
-  let override = {
-    nonce: nonce,
-  };
-  let response = await writeContract.summon(type, override);
-  let receipt = await response.wait();
-  let id = receipt.events[1].args.summoner.toNumber();
-  console.log(id);
-  pause();
+    let nonce = await wallet.getTransactionCount();
+    let override = {
+        nonce: nonce,
+    };
+    let response = await writeContract.summon(type, override);
+    let receipt = await response.wait();
+    let id = receipt.events[1].args.summoner.toNumber();
+    console.log(id);
+    pause();
 };
 
 const main = async () => {
-  for (let b = 0; b < 5; b++) {
-    for (let i = 1; i < 12; i++) {
-      await mint(i);
+    for (let b = 0; b < 5; b++) {
+        for (let i = 1; i < 12; i++) {
+            await mint(i);
+        }
     }
-  }
 };
 
-main().catch((err) => {
-  console.error("there was an error");
+main().catch(() => {
+    console.error("there was an error");
 });
 
 // mint(1).catch((err) => {
