+++ config/contracts.js
@@ -1,8 +1,8 @@
 const contracts = {
-  rarity: "0xce761D788DF608BD21bdd59d6f4B54b2e27F25Bb",
-  gold: "0x2069B76Afe6b734Fb65D1d099E7ec64ee9CC76B2",
-  attributes: "0xB5F5AF1087A8DA62A23b08C00C6ec9af21F397a1",
-  crafting1_1: "0x2A0F1cB17680161cF255348dDFDeE94ea8Ca196A",
+    rarity: "0xce761D788DF608BD21bdd59d6f4B54b2e27F25Bb",
+    gold: "0x2069B76Afe6b734Fb65D1d099E7ec64ee9CC76B2",
+    attributes: "0xB5F5AF1087A8DA62A23b08C00C6ec9af21F397a1",
+    crafting1_1: "0x2A0F1cB17680161cF255348dDFDeE94ea8Ca196A",
 };
 
 module.exports = contracts;
