+++ main.js
@@ -5,24 +5,25 @@ const summonerIds = require("./actions/summoners");
 const ethers = require("ethers");
 const { provider } = require("./config/wallet");
 const { checkClass } = require("./actions/classes");
+const { spendBaseAttributes } = require("./actions/spendBaseAttributes");
 
 const main = async () => {
-  let block = await provider.getBlock();
-  let currentTime = ethers.BigNumber.from(block.timestamp);
+    let block = await provider.getBlock();
+    let currentTime = ethers.BigNumber.from(block.timestamp);
 
-  for (let i = 0; i < summonerIds.length; i++) {
-    let summonerClass = await checkClass(summonerIds[i]);
-    console.log(
-      `### Start with summoner ${summonerIds[i]} ${summonerClass} ###`
-    );
-    await adventure(summonerIds[i], currentTime);
-    await levelUp(summonerIds[i]);
-    await craftAdventure(summonerIds[i], currentTime);
-    await pointBuy(summonerIds[i]);
-    console.log(``);
-  }
+    for (let i = 0; i < summonerIds.length; i++) {
+        let summonerClass = await checkClass(summonerIds[i]);
+        console.log(
+            `### Start with summoner ${summonerIds[i]} ${summonerClass} ###`
+        );
+        await adventure(summonerIds[i], currentTime);
+        await levelUp(summonerIds[i]);
+        await craftAdventure(summonerIds[i], currentTime);
+        await spendBaseAttributes(summonerIds[i]);
+        console.log(``);
+    }
 };
 
 main().catch((err) => {
-  console.error(err);
+    console.error(err);
 });
